Merging our implementation into gitflow
Credit to
https://github.com/marcricharme/gitflow
Where this work is deriving from.

Enabling git-flow to be used with gerrit,
Requires some special access controls in gerrit for the scope refs/heads/feature
To be able to fully use this fork.
Access controls like this.
[access "refs/heads/feature/*"]
	create = group developers
	push = +force group developers
	pushMerge = group developers
	label-Code-Review = -2..+2 group developers
	label-Verified = -1..+1 group developers
	submit = group developers
	forgeAuthor = group developers

If you don't like this in the feature scope
the function gitflow_configure_prefixes() in gitflow-common will set the prefix used in the script
so you can change it to whatever you like